feat(welcome): move roadmap to layout-kit grouped-list, drop washi-tape
Colocates section-roadmap under its own dir, swaps the hand-rolled list
for grouped-list, and removes the washi-tape image accent.
