+++ src/views/welcome/index.vue
@@ -7,8 +7,8 @@ import { provideWelcomeLayout } from './welcome-layout'
 import Splash from './splash/index.vue'
 import SectionFeatures from './section-features/index.vue'
 import SectionPricing from './section-pricing/index.vue'
-import SectionRoadmap from './section-roadmap.vue'
-import WelcomeFooter from '@/components/welcome-footer.vue'
+import SectionRoadmap from './section-roadmap/index.vue'
+import WelcomeFooter from '@/views/welcome/welcome-footer.vue'
 
 const session = useSessionStore()
 const { open: openSignup } = useSignupModal()
